refactor(session): v0.0.1
Refactor and fixes code after session controller implementation

+++ application/entrypoint/src/main/kotlin/br/com/fairie/partypay/endpoints/menu/mapper/MenuDtoMapper.kt
@@ -37,5 +37,5 @@ fun Order.toDto(): OrderDTO =
     OrderDTO(
         name = name,
         description = description,
-        value = value
+        value = value.toDouble()
     )
